+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 0.8.2
+
 ### Fixed
 - ProtoLanguageFileWriter now uses the real filename instead of
   inferring it from the package name.